+++ CHANGELOG.rst
@@ -16,6 +16,11 @@ CHANGELOG
 - More detailed information in ValidationErrors
   [masipcat]
 
+- Provide way to configure content types as not globally addable
+  [lferran]
+
+- Fix Users and Groups to be addable only on manager folders [lferran]
+
 - Fix optimized lookup to work with fields that do not have `_type`
   [vangheem]